Architecture changed, player module code shifted to Game component instead of App file
